pyproject.toml
[project]
name = "gitlab-hooks"
version = "1.0.0"
description = "API for handling GitLab webhooks"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-decouple>=3.8",
    "httpx>=0.25.0",
    "motor>=3.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
dev = "dev:main"

